split service/implementation
