+++ backend/typescript/services/interfaces/IEntityService.ts
@@ -1,61 +1,26 @@
-export interface EntityRequestDTO {
-  stringField: string;
-  intField: number;
-  enumField: string;
-  stringArrayField: [string];
-  boolField: boolean;
+import { Entity, IEntity } from "../../models/entity.mgmodel";
+
+enum Options {
+    A, B, C, D,
 }
 
-export interface EntityResponseDTO {
-  id: string;
-  stringField: string;
-  intField: number;
-  enumField: string;
-  stringArrayField: [string];
-  boolField: boolean;
+export interface EntityRequestDTO {
+    string_field: string;
+    int_field: number;
+    enum_field: string;
+    string_array_field: [string];
+    bool_field: boolean;
 }
 
 export interface IEntityService {
-  /**
-   * retrieve the Entity with the given id
-   * @param id entity id
-   * @returns requested Entity
-   * @throws Error if retrieval fails
-   */
-  getEntity(id: string): Promise<EntityResponseDTO>;
-
-  /**
-   * retrieve all Entities
-   * @param
-   * @returns returns array of Entities
-   * @throws Error if retrieval fails
-   */
-  getEntities(): Promise<EntityResponseDTO[]>;
-
-  /**
-   * create an Entity with the fields given in the DTO, return created Entity
-   * @param entity user's email
-   * @returns the created Entity
-   * @throws Error if creation fails
-   */
-  createEntity(entity: EntityRequestDTO): Promise<EntityResponseDTO>;
-
-  /**
-   * update the Entity with the given id with fields in the DTO, return updated Entity
-   * @param id entity id
-   * @param entity Updated Entity
-   * @returns the updated Entity
-   * @throws Error if update fails
-   */
-  updateEntity(
-    id: string,
-    entity: EntityRequestDTO,
-  ): Promise<EntityResponseDTO | null>;
-
-  /**
-   * delete the entity with the given id
-   * @param id entity id
-   * @throws Error if deletion fails
-   */
-  deleteEntity(id: string): Promise<void>;
+    /* retrieve the Entity with the given id */
+    getEntity(id: string): Promise<IEntity>
+    /* retrieve all Entities (pagination is nice-to-have future feature) */
+    getEntities(): Promise<IEntity[]>
+    /* create an Entity with the fields given in the DTO, return created Entity */
+    createEntity(entity: IEntity): Promise<IEntity>;
+    /* update the Entity with the given id with fields in the DTO, return updated Entity */
+    updateEntity(id: string, entity: IEntity): Promise<IEntity | null>;
+    /* delete the entity with the given id */
+    deleteEntity(id: string): void;
 }
